requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.32
psycopg[binary]>=3.2
jinja2>=3.1
pyyaml>=6.0
markdown>=3.7
pymdown-extensions>=10.0
